requirement.txt
google-generativeai==0.8.4
pillow==11.1.0
streamlit==1.41.1
streamlit-option-menu==0.4.0
python-dotenv==1.0.1
groq==0.9.0
youtube_transcript_api
pathlib